rename providers-installation->providers_installation
