+++ controllers/dbaasplatform_controller.go
@@ -27,7 +27,7 @@ import (
 	dbaasv1alpha1 "github.com/RHEcosystemAppEng/dbaas-operator/api/v1alpha1"
 	"github.com/RHEcosystemAppEng/dbaas-operator/controllers/reconcilers"
 	"github.com/RHEcosystemAppEng/dbaas-operator/controllers/reconcilers/console_plugin"
-	providers_installation "github.com/RHEcosystemAppEng/dbaas-operator/controllers/reconcilers/providers-installation"
+	providers_installation "github.com/RHEcosystemAppEng/dbaas-operator/controllers/reconcilers/providers_installation"
 	"github.com/RHEcosystemAppEng/dbaas-operator/controllers/reconcilers/quickstart_installation"
 
 	"github.com/go-logr/logr"
